Moving the scripts into the CUDLR namespace Moving Console to CUDLR.Console Moving ConsoleServer to CUDLR.Server Moving ConsoleCommand to CUDLR.Command Moving ConsoleRoute to CUDLR.Route Cleaning up static file types in CUDLR.Server
